requirements/base.txt
pandas>=2.2
numpy>=1.26
pyyaml>=6.0
pydantic>=2.7
typer[all]>=0.12
python-dateutil>=2.9
exchange-calendars>=4.5
pyarrow>=17.0
rich>=13.7
tqdm>=4.66
matplotlib>=3.9
scipy>=1.13
networkx>=3.3
yfinance>=0.2.27
PySide6==6.8.0
shiboken6==6.8.0
opencv-python-headless==4.10.*
pyqtdarktheme==2.1.0

